Fix AppVeyor scripts not exiting properly when an error occurred

+++ appveyor/PackageArtifacts.ps1
@@ -11,8 +11,12 @@ if ($env:PLATFORM -eq "x86" -or $env:PLATFORM -eq "x64") {
     Move-Item artifacts\work\bin\LightFX.dll artifacts\work\x64\
 
     7z a -r "artifacts\LightFX-Extender-x86_v$($env:APPVEYOR_BUILD_VERSION).zip" "$($env:APPVEYOR_BUILD_FOLDER)\artifacts\work\x86\*.*"
+    if ($LastExitCode -ne 0) exit $LastExitCode
     7z a -r "artifacts\LightFX-Extender-x64_v$($env:APPVEYOR_BUILD_VERSION).zip" "$($env:APPVEYOR_BUILD_FOLDER)\artifacts\work\x64\*.*"
+    if ($LastExitCode -ne 0) exit $LastExitCode
     7z a -r "artifacts\LightFX-Extender-full_v$($env:APPVEYOR_BUILD_VERSION).zip" "$($env:APPVEYOR_BUILD_FOLDER)\artifacts\work\*.*"
+    if ($LastExitCode -ne 0) exit $LastExitCode
     ISCC appveyor\BuildInstaller.iss
+    if ($LastExitCode -ne 0) exit $LastExitCode
     Remove-Item -Recurse -Force artifacts\work
 }
